backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
httpx
pydantic
motor
aiohttp 
pyjwt
